Support HTTP 307 and HTTP 308 redirects
This commit provides support for following HTTP 307 and HTTP 308
redirects. Specifically, the `URL.openRedirectingConnection` function
(located in `GenericResponse.kt`) was updated to check for the HTTP
response code's presence in an array of redirect response codes
(301, 302, 303, 307, and 308). Previously, the function was only
checking for response codes in the range 301..303.

For information regarding these additional HTTP status codes, see:
  - [307 Temporary Redirect](https://tools.ietf.org/html/rfc7231#section-6.4.7)
  - [308 Permanent Redirect](https://tools.ietf.org/html/rfc7538#section-3)